--- tests/unittests/code-IsMailAddress_20160427.cpp
@@ -1,76 +1,0 @@
-﻿// テスト対象の旧関数(コピペで埋め込み)
-#include "StdAfx.h"
-
-/* 現在位置がメールアドレスならば、NULL以外と、その長さを返す
-	@date 2016.04.27 記号類を許可
-*/
-BOOL IsMailAddress_20160427( const wchar_t* pszBuf, int nBufLen, int* pnAddressLenfth )
-{
-	int		j;
-	int		nDotCount;
-	int		nBgn;
-
-	j = 0;
-	if( (pszBuf[j] >= L'a' && pszBuf[j] <= L'z')
-	 || (pszBuf[j] >= L'A' && pszBuf[j] <= L'Z')
-	 || (pszBuf[j] >= L'0' && pszBuf[j] <= L'9')
-	 || NULL != wcschr(L"!#$%&'*+-/=?^_`{|}~", pszBuf[j])
-	){
-		j++;
-	}else{
-		return FALSE;
-	}
-	while( j < nBufLen - 2 &&
-		(
-		(pszBuf[j] >= L'a' && pszBuf[j] <= L'z')
-	 || (pszBuf[j] >= L'A' && pszBuf[j] <= L'Z')
-	 || (pszBuf[j] >= L'0' && pszBuf[j] <= L'9')
-	 || (pszBuf[j] == L'.')
-	 || NULL != wcschr(L"!#$%&'*+-/=?^_`{|}~", pszBuf[j])
-		)
-	){
-		j++;
-	}
-	if( j == 0 || j >= nBufLen - 2  ){
-		return FALSE;
-	}
-	if( L'@' != pszBuf[j] ){
-		return FALSE;
-	}
-//	nAtPos = j;
-	j++;
-	nDotCount = 0;
-//	nAlphaCount = 0;
-
-	for (;;) {
-		nBgn = j;
-		while( j < nBufLen &&
-			(
-			(pszBuf[j] >= L'a' && pszBuf[j] <= L'z')
-		 || (pszBuf[j] >= L'A' && pszBuf[j] <= L'Z')
-		 || (pszBuf[j] >= L'0' && pszBuf[j] <= L'9')
-		 || (pszBuf[j] == L'-')
-		 || (pszBuf[j] == L'_')
-			)
-		){
-			j++;
-		}
-		if( 0 == j - nBgn ){
-			return FALSE;
-		}
-		if( L'.' != pszBuf[j] ){
-			if( 0 == nDotCount ){
-				return FALSE;
-			}else{
-				break;
-			}
-		}else{
-			nDotCount++;
-			j++;
-		}
-	}
-	if( NULL != pnAddressLenfth ){
-		*pnAddressLenfth = j;
-	}
-	return TRUE;
-}
